Moved code to profilecreator pkg and added unit tests
