+++ cmd/performance-profile-creator/cmd/root.go
@@ -18,40 +18,18 @@ package cmd
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
-	"path"
-	"path/filepath"
-	"time"
 
+	"github.com/openshift-kni/performance-addon-operators/pkg/profilecreator"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/spf13/cobra"
 	yaml "gopkg.in/yaml.v2"
-	k8syaml "k8s.io/apimachinery/pkg/util/yaml"
-	"k8s.io/component-helpers/scheduling/corev1"
 
 	"k8s.io/utils/pointer"
 
 	performancev2 "github.com/openshift-kni/performance-addon-operators/api/v2"
-	machineconfigv1 "github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1"
-	v1 "k8s.io/api/core/v1"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-const (
-	// ClusterScopedResources defines the subpath, relative to the top-level must-gather directory.
-	// A top-level must-gather directory is of the following format:
-	// must-gather-dir/quay-io-openshift-kni-performance-addon-operator-must-gather-sha256-<Image SHA>
-	// Here we find the cluster-scoped definitions saved by must-gather
-	ClusterScopedResources = "cluster-scoped-resources"
-	// CoreNodes defines the subpath, relative to ClusterScopedResources, on which we find node-specific data
-	CoreNodes = "core/nodes"
-	// MCPools defines the subpath, relative to ClusterScopedResources, on which we find the machine config pool definitions
-	MCPools = "machineconfiguration.openshift.io/machineconfigpools"
-	// YAMLSuffix is the extension of the yaml files saved by must-gather
-	YAMLSuffix = ".yaml"
 )
 
 // rootCmd represents the base command when called without any subcommands
@@ -61,23 +39,19 @@ var rootCmd = &cobra.Command{
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		mcpName := cmd.Flag("mcp-name").Value.String()
 		mustGatherDirPath := cmd.Flag("must-gather-dir-path").Value.String()
-		mcp, err := getMCP(mustGatherDirPath, mcpName)
+		mcp, err := profilecreator.GetMCP(mustGatherDirPath, mcpName)
 		if err != nil {
 			return fmt.Errorf("Error obtaining MachineConfigPool %s: %v", mcpName, err)
 		}
 		labelSelector := mcp.Spec.NodeSelector
-		nodes, err := getNodeList(mustGatherDirPath)
+		nodes, err := profilecreator.GetNodeList(mustGatherDirPath)
 		if err != nil {
 			return fmt.Errorf("Error obtaining Nods %s: %v", mcpName, err)
 		}
 
-		matchedNodes := make([]*v1.Node, 0)
-		for _, node := range nodes {
-			matches, _ := corev1.MatchNodeSelectorTerms(node, getNodeSelectorFromLabelSelector(labelSelector))
-			if matches {
-				log.Infof("%s is targetted by %s MCP", node.GetName(), mcpName)
-				matchedNodes = append(matchedNodes, node)
-			}
+		matchedNodes, err := profilecreator.GetMatchedNodes(nodes, labelSelector)
+		for _, node := range matchedNodes {
+			log.Infof("%s is targetted by %s MCP", node.GetName(), mcpName)
 		}
 		return nil
 	},
@@ -85,129 +59,6 @@ var rootCmd = &cobra.Command{
 		profileName := cmd.Flag("profile-name").Value.String()
 		createProfile(profileName)
 	},
-}
-
-func getNodeSelectorFromLabelSelector(labelSelector *metav1.LabelSelector) *v1.NodeSelector {
-
-	matchExpressions := make([]v1.NodeSelectorRequirement, 0)
-	for key, value := range labelSelector.MatchLabels {
-		matchExpressions = append(matchExpressions, v1.NodeSelectorRequirement{
-			Key:      key,
-			Operator: v1.NodeSelectorOpIn,
-			Values:   []string{value},
-		})
-	}
-	matchFields := make([]v1.NodeSelectorRequirement, 0)
-	for _, labelSelectorRequirement := range labelSelector.MatchExpressions {
-		matchExpressions = append(matchFields, v1.NodeSelectorRequirement{
-			Key:      labelSelectorRequirement.Key,
-			Operator: v1.NodeSelectorOperator(string(labelSelectorRequirement.Operator)),
-			Values:   labelSelectorRequirement.Values,
-		})
-	}
-
-	nodeSelectorTerms := []v1.NodeSelectorTerm{
-		{
-			MatchExpressions: matchExpressions,
-			MatchFields:      matchFields,
-		},
-	}
-	nodeSelector := &v1.NodeSelector{
-		NodeSelectorTerms: nodeSelectorTerms,
-	}
-
-	return nodeSelector
-
-}
-
-func getMustGatherFullPaths(mustGatherPath string, suffix string) (string, error) {
-	// The glob pattern below depends on the must gather image name. It is assumed here
-	// that the image would have "performance-addon-operator-must-gather" substring in the name.
-	paths, err := filepath.Glob(mustGatherPath + "/*performance-addon-operator-must-gather*/" + suffix)
-
-	// Out of all the paths that match the glob pattern we choose the path that was modified most recently.
-	var latestTime time.Time
-	var lastModifiedPath string
-
-	for _, path := range paths {
-		info, err := os.Stat(path)
-		if err != nil {
-			return "", fmt.Errorf("Error obtaining the path stats %s: %v", path, err)
-		}
-		if info.ModTime().After(latestTime) {
-			lastModifiedPath = path
-		}
-	}
-	if len(paths) > 1 {
-		log.Infof("Multiple matches for the specified must gather directory path: %s and suffix: %s", mustGatherPath, suffix)
-		log.Infof("Selecting the most fresh path (the path that was last modified): %s", lastModifiedPath)
-	}
-	return lastModifiedPath, err
-}
-
-func getNode(mustGatherDirPath, nodeName string) (*v1.Node, error) {
-	var node v1.Node
-	nodePathSuffix := path.Join(ClusterScopedResources, CoreNodes, nodeName)
-	path, err := getMustGatherFullPaths(mustGatherDirPath, nodePathSuffix)
-	if err != nil {
-		return nil, fmt.Errorf("Error obtaining MachineConfigPool %s: %v", nodeName, err)
-	}
-
-	src, err := os.Open(path)
-	if err != nil {
-		return nil, fmt.Errorf("Error opening %q: %v", path, err)
-	}
-	defer src.Close()
-
-	dec := k8syaml.NewYAMLOrJSONDecoder(src, 1024)
-	if err := dec.Decode(&node); err != nil {
-		return nil, fmt.Errorf("Error opening %q: %v", path, err)
-	}
-	return &node, nil
-}
-
-func getNodeList(mustGatherDirPath string) ([]*v1.Node, error) {
-	machines := make([]*v1.Node, 0)
-
-	nodePathSuffix := path.Join(ClusterScopedResources, CoreNodes)
-	nodePath, err := getMustGatherFullPaths(mustGatherDirPath, nodePathSuffix)
-	if err != nil {
-		return nil, fmt.Errorf("Error obtaining Nodes: %v", err)
-	}
-	nodes, err := ioutil.ReadDir(nodePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list mustGatherPath directories: %v", err)
-	}
-	for _, node := range nodes {
-		nodeName := node.Name()
-		node, err := getNode(mustGatherDirPath, nodeName)
-		if err != nil {
-			return nil, fmt.Errorf("Error obtaining Nodes %s: %v", nodeName, err)
-		}
-		machines = append(machines, node)
-	}
-	return machines, nil
-}
-
-func getMCP(mustGatherDirPath, mcpName string) (*machineconfigv1.MachineConfigPool, error) {
-	var mcp machineconfigv1.MachineConfigPool
-
-	mcpPathSuffix := path.Join(ClusterScopedResources, MCPools, mcpName+YAMLSuffix)
-	mcpPath, err := getMustGatherFullPaths(mustGatherDirPath, mcpPathSuffix)
-	if err != nil {
-		return nil, fmt.Errorf("Error obtaining MachineConfigPool %s: %v", mcpName, err)
-	}
-
-	src, err := os.Open(mcpPath)
-	if err != nil {
-		return nil, fmt.Errorf("Error opening %q: %v", mcpPath, err)
-	}
-	defer src.Close()
-	dec := k8syaml.NewYAMLOrJSONDecoder(src, 1024)
-	if err := dec.Decode(&mcp); err != nil {
-		return nil, fmt.Errorf("Error opening %q: %v", mcpPath, err)
-	}
-	return &mcp, nil
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
