+++ pkg/profilecreator/profilecreator_suite_test.go
@@ -1,0 +1,13 @@
+package profilecreator
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestProfileCreator(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Profile Creator Suite")
+}
