+++ testdata/must-gather/must-gather.local.directory/quay-io-openshift-kni-performance-addon-operator-must-gather-sha256-09d31edb2171e1fe385df1238ce5a6f7ca3556efa271550a7e98e26dce290128/cluster-scoped-resources/machineconfiguration.openshift.io/machineconfigpools/master.yaml
@@ -1,0 +1,140 @@
+apiVersion: machineconfiguration.openshift.io/v1
+kind: MachineConfigPool
+metadata:
+  creationTimestamp: "2020-11-25T07:59:09Z"
+  generation: 2
+  labels:
+    machineconfiguration.openshift.io/mco-built-in: ""
+    operator.machineconfiguration.openshift.io/required-for-upgrade: ""
+    pools.operator.machineconfiguration.openshift.io/master: ""
+  managedFields:
+  - apiVersion: machineconfiguration.openshift.io/v1
+    fieldsType: FieldsV1
+    fieldsV1:
+      f:metadata:
+        f:labels:
+          .: {}
+          f:machineconfiguration.openshift.io/mco-built-in: {}
+          f:operator.machineconfiguration.openshift.io/required-for-upgrade: {}
+          f:pools.operator.machineconfiguration.openshift.io/master: {}
+      f:spec:
+        .: {}
+        f:configuration: {}
+        f:machineConfigSelector:
+          .: {}
+          f:matchLabels:
+            .: {}
+            f:machineconfiguration.openshift.io/role: {}
+        f:nodeSelector:
+          .: {}
+          f:matchLabels:
+            .: {}
+            f:node-role.kubernetes.io/master: {}
+        f:paused: {}
+    manager: machine-config-operator
+    operation: Update
+    time: "2020-11-25T07:59:09Z"
+  - apiVersion: machineconfiguration.openshift.io/v1
+    fieldsType: FieldsV1
+    fieldsV1:
+      f:spec:
+        f:configuration:
+          f:name: {}
+          f:source: {}
+      f:status:
+        .: {}
+        f:conditions: {}
+        f:configuration:
+          .: {}
+          f:name: {}
+          f:source: {}
+        f:degradedMachineCount: {}
+        f:machineCount: {}
+        f:observedGeneration: {}
+        f:readyMachineCount: {}
+        f:unavailableMachineCount: {}
+        f:updatedMachineCount: {}
+    manager: machine-config-controller
+    operation: Update
+    time: "2020-11-25T08:01:39Z"
+  name: master
+  resourceVersion: "11223"
+  selfLink: /apis/machineconfiguration.openshift.io/v1/machineconfigpools/master
+  uid: 8589c30f-ea32-49cb-8bf8-4ba9dad44856
+spec:
+  configuration:
+    name: rendered-master-e92b311a0208749dbba5c4458afcc653
+    source:
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 00-master
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 01-master-container-runtime
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 01-master-kubelet
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 99-master-generated-registries
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 99-master-ssh
+  machineConfigSelector:
+    matchLabels:
+      machineconfiguration.openshift.io/role: master
+  nodeSelector:
+    matchLabels:
+      node-role.kubernetes.io/master: ""
+  paused: false
+status:
+  conditions:
+  - lastTransitionTime: "2020-11-25T08:00:37Z"
+    message: ""
+    reason: ""
+    status: "False"
+    type: RenderDegraded
+  - lastTransitionTime: "2020-11-25T08:01:39Z"
+    message: All nodes are updated with rendered-master-e92b311a0208749dbba5c4458afcc653
+    reason: ""
+    status: "True"
+    type: Updated
+  - lastTransitionTime: "2020-11-25T08:01:39Z"
+    message: ""
+    reason: ""
+    status: "False"
+    type: Updating
+  - lastTransitionTime: "2020-11-25T08:01:39Z"
+    message: ""
+    reason: ""
+    status: "False"
+    type: NodeDegraded
+  - lastTransitionTime: "2020-11-25T08:01:39Z"
+    message: ""
+    reason: ""
+    status: "False"
+    type: Degraded
+  configuration:
+    name: rendered-master-e92b311a0208749dbba5c4458afcc653
+    source:
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 00-master
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 01-master-container-runtime
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 01-master-kubelet
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 99-master-generated-registries
+    - apiVersion: machineconfiguration.openshift.io/v1
+      kind: MachineConfig
+      name: 99-master-ssh
+  degradedMachineCount: 0
+  machineCount: 3
+  observedGeneration: 2
+  readyMachineCount: 3
+  unavailableMachineCount: 0
+  updatedMachineCount: 3
